Updated to work with custom Transform class

+++ src/LandMark.cpp
@@ -8,31 +8,31 @@
 #include "LandMark.h"
 #include <tuple>
 
-LandMark::LandMark(double x, double y, double z)
+LandMark::LandMark(float x, float y, float z)
 {
     LandMark::_x = x;
     LandMark::_y = y;
     LandMark::_z = z;
 }
 
-void LandMark::SetTranslationalVector(double x, double y, double z)
+void LandMark::SetTranslationalVector(float x, float y, float z)
 {
     LandMark::_x = x;
     LandMark::_y = y;
     LandMark::_z = z;
 }
 
-std::tuple<double, double, double> LandMark::GetTranslationalVector()
+std::tuple<float, float, float> LandMark::GetTranslationalVector()
 {
-    return std::tuple<double, double, double>{_x, _y, _z};
+    return std::tuple<float, float, float>{_x, _y, _z};
 }
 
 // int main(){
 //     LandMark newLM = LandMark(0, 0, 0);
 //     std::tuple tLM = newLM.GetTranslationalVector();
-//     double xLM = std::get<0>(tLM);
-//     double yLM = std::get<1>(tLM);
-//     double zLM = std::get<2>(tLM);
+//     float xLM = std::get<0>(tLM);
+//     float yLM = std::get<1>(tLM);
+//     float zLM = std::get<2>(tLM);
 //     std::cout << "My LM translational vector: (" << xLM << ", " << yLM << ", " << zLM << ")" << std::endl;
 //     return 0;
 // }
